+++ packages/happn-cluster/test/unit/services/04-unit-utils.js
@@ -1,6 +1,4 @@
-require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
-  const NODE_MAJOR_VERSION = process.versions.node.split('.')[0];
-  const ip4Family = NODE_MAJOR_VERSION < 18 ? 'IPv4' : 4;
+require('../../lib/test-helper').describe({ timeout: 120e3 }, function (test) {
   it('tests get-address', () => {
     let logs = [];
     let mockLogger = {
@@ -17,15 +15,87 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
         getAddress({
           eth1: [
             {
-              address: '169.254.0.1',
+              address: '190.254.0.1',
+              family: 'IPv4',
             },
             {
-              address: '169.254.0.2',
+              address: '190.254.0.2',
+              family: 'IPv4',
             },
           ],
         })
       )
-      .to.eql('169.254.0.2');
+      .to.eql('190.254.0.2');
+
+    test
+      .expect(
+        getAddress({
+          eth0: [
+            {
+              address: '192.254.1.1',
+              family: 'IPv4',
+            },
+          ],
+          eth1: [
+            {
+              address: '192.254.0.1', // this gets selected because the item on index 1 is IPv6
+              family: 'IPv4',
+            },
+            {
+              address: '192.254.0.2',
+              family: 'IPv6',
+            },
+          ],
+        })
+      )
+      .to.eql('192.254.0.1');
+
+    test
+      .expect(
+        getAddress({
+          eth0: [
+            {
+              address: '192.254.1.1', // not selected because not eth1
+              family: 'IPv4',
+            },
+          ],
+          eth1: [
+            {
+              address: '192.254.0.1', // this gets selected because the item on index 1 is a system address
+              family: 'IPv4',
+            },
+            {
+              address: '169.254.0.2',
+              family: 'IPv4',
+            },
+          ],
+        })
+      )
+      .to.eql('192.254.0.1');
+
+    test
+      .expect(
+        getAddress({
+          eth0: [
+            {
+              address: '192.254.1.1', // not selected because not eth1
+              family: 'IPv4',
+            },
+          ],
+          eth1: [
+            {
+              address: '192.254.0.1', // this gets selected because the item on index 1 is marked as internal
+              family: 'IPv4',
+            },
+            {
+              address: '129.254.0.2',
+              family: 'IPv4',
+              internal: true,
+            },
+          ],
+        })
+      )
+      .to.eql('192.254.0.1');
 
     getAddress = testGetAddress(mockLogger, {
       NETWORK_INTERFACE_ID: 'eth1',
@@ -36,15 +106,17 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
         getAddress({
           eth1: [
             {
-              address: '169.254.0.1',
+              address: '190.254.0.1',
+              family: 'IPv4',
             },
             {
-              address: '169.254.0.2',
+              address: '190.254.0.2',
+              family: 'IPv4',
             },
           ],
         })
       )
-      .to.eql('169.254.0.1');
+      .to.eql('190.254.0.1');
 
     getAddress = testGetAddress(
       mockLogger,
@@ -57,10 +129,12 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
           return {
             eth1: [
               {
-                address: '169.254.0.3',
+                address: '190.254.0.3',
+                family: 'IPv4',
               },
               {
-                address: '169.254.0.4',
+                address: '190.254.0.4',
+                family: 'IPv4',
               },
             ],
           };
@@ -68,7 +142,7 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
       }
     );
 
-    test.expect(getAddress()).to.eql('169.254.0.3');
+    test.expect(getAddress()).to.eql('190.254.0.3');
 
     getAddress = testGetAddress(
       mockLogger,
@@ -81,10 +155,12 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
           return {
             eth1: [
               {
-                address: '169.254.0.3',
+                address: '190.254.0.3',
+                family: 'IPv4',
               },
               {
-                address: '169.254.0.4',
+                address: '190.254.0.4',
+                family: 'IPv4',
               },
             ],
           };
@@ -92,7 +168,7 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
       }
     );
 
-    test.expect(getAddress()).to.eql('169.254.0.4');
+    test.expect(getAddress()).to.eql('190.254.0.4');
 
     getAddress = testGetAddress(
       mockLogger,
@@ -106,9 +182,11 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
             blah: [
               {
                 address: '169.254.0.3',
+                family: 'IPv4',
               },
               {
                 address: '169.254.0.4',
+                family: 'IPv4',
               },
             ],
           };
@@ -139,9 +217,11 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
             blah: [
               {
                 address: '169.254.0.3',
+                family: 'IPv4',
               },
               {
                 address: '169.254.0.4',
+                family: 'IPv4',
               },
             ],
           };
@@ -171,9 +251,11 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
             eth1: [
               {
                 address: '169.254.0.3',
+                family: 'IPv4',
               },
               {
                 address: '169.254.0.4',
+                family: 'IPv4',
               },
             ],
           };
@@ -195,7 +277,7 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
       (testInterface, interfaceKey) => {
         let found = interfaces[interfaceKey];
         found.forEach((interfaceItem, interfaceItemIndex) => {
-          if (!interfaceItem.internal && interfaceItem.family === ip4Family) {
+          if (!interfaceItem.internal && [4, 'IPv4'].includes(interfaceItem.family)) {
             testInterface = {
               id: interfaceKey,
               index: interfaceItemIndex,
@@ -215,6 +297,105 @@ require('../../lib/test-helper').describe({ timeout: 40e3 }, function (test) {
 
     test.expect(getAddress()).to.eql(testInterface.address);
   });
+
+  it('tests get-address - defaults', () => {
+    testAddressCorrect('eth0', 'ens33');
+    testAddressCorrect('ens33', 'eth0');
+    testAddressCorrect('en0', 'eth0');
+  });
+
+  it('tests log warning', () => {
+    let mockLogger = {
+      warn: test.sinon.stub(),
+    };
+    let getAddress = testGetAddress(mockLogger, {
+      NETWORK_INTERFACE_ID: 'eth1',
+      NETWORK_INTERFACE: 1,
+    });
+    getAddress({
+      eth2: [
+        {
+          address: '190.254.0.3',
+          family: 'IPv4',
+        },
+        {
+          address: '190.254.0.4',
+          family: 'IPv4',
+        },
+      ],
+    });
+    test
+      .expect(mockLogger.warn.lastCall.args[0])
+      .to.be(
+        'get address for SWIM or cluster: interface with id [eth1] not found or address index out of bounds - dynamically resolved to address [190.254.0.3] on NIC [eth2]'
+      );
+    getAddress = testGetAddress(mockLogger, {
+      NETWORK_INTERFACE_ID: 'eth2',
+      NETWORK_INTERFACE: 3,
+    });
+    getAddress({
+      eth2: [
+        {
+          address: '190.254.0.3',
+          family: 'IPv4',
+        },
+        {
+          address: '190.254.0.4',
+          family: 'IPv4',
+        },
+      ],
+    });
+    test
+      .expect(mockLogger.warn.lastCall.args[0])
+      .to.be(
+        'get address for SWIM or cluster: interface with id [eth2] not found or address index out of bounds - dynamically resolved to address [190.254.0.3] on NIC [eth2]'
+      );
+  });
+
+  it('covers just running as standard', () => {
+    let address = require('../../../lib/utils/get-address')()();
+    const regexExp =
+      /^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])$/gi;
+    test.expect(regexExp.test(address)).to.be(true);
+  });
+
+  function testAddressCorrect(specifiedInterfaceId, badInterfaceId) {
+    let logs = [];
+    let mockLogger = {
+      warn: (msg) => {
+        logs.push(msg);
+      },
+    };
+    let address = '191.120.100.100';
+    let mockConfiguration = [
+      mockLogger,
+      {
+        NETWORK_INTERFACE_ID: specifiedInterfaceId,
+      },
+      {
+        networkInterfaces: () => {
+          return {
+            [badInterfaceId]: [
+              {
+                address: 'bad',
+                family: 4,
+              },
+            ],
+            [specifiedInterfaceId]: [
+              {
+                address,
+                family: 4,
+              },
+            ],
+          };
+        },
+      },
+    ];
+    const getAddress = testGetAddress(...mockConfiguration);
+    test
+      .expect(getAddress())
+      .to.eql(mockConfiguration[2].networkInterfaces()[specifiedInterfaceId][0].address);
+  }
 
   function testGetAddress(logger, env, os) {
     return require('../../../lib/utils/get-address')(logger, env, os);